+++ src/main/java/org/truffleruby/parser/parser/ParserSupport.java
@@ -1224,8 +1224,8 @@ public class ParserSupport {
         this.lexer = lexer;
     }
 
-    public DStrParseNode createDStrNode(SourceIndexLength position, boolean frozen) {
-        return new DStrParseNode(position, lexer.getEncoding(), frozen);
+    public DStrParseNode createDStrNode(SourceIndexLength position) {
+        return new DStrParseNode(position, lexer.getEncoding());
     }
 
     public ParseNodeTuple createKeyValue(ParseNode key, ParseNode value) {
@@ -1281,12 +1281,12 @@ public class ParserSupport {
         }
 
         if (head instanceof EvStrParseNode) {
-            head = createDStrNode(head.getPosition(), false).add(head);
+            head = createDStrNode(head.getPosition()).add(head);
         }
 
         if (lexer.getHeredocIndent() > 0) {
             if (head instanceof StrParseNode) {
-                head = createDStrNode(head.getPosition(), false).add(head);
+                head = createDStrNode(head.getPosition()).add(head);
                 return list_append(head, tail);
             } else if (head instanceof DStrParseNode) {
                 return list_append(head, tail);
@@ -1309,10 +1309,7 @@ public class ParserSupport {
 
         } else if (tail instanceof DStrParseNode) {
             if (head instanceof StrParseNode) { // Str + oDStr -> Dstr(Str, oDStr.contents)
-                DStrParseNode newDStr = new DStrParseNode(
-                        head.getPosition(),
-                        ((DStrParseNode) tail).getEncoding(),
-                        false);
+                DStrParseNode newDStr = new DStrParseNode(head.getPosition(), ((DStrParseNode) tail).getEncoding());
                 newDStr.add(head);
                 newDStr.add(tail);
                 return newDStr;
@@ -1323,15 +1320,12 @@ public class ParserSupport {
 
         // tail must be EvStrParseNode at this point
         if (head instanceof StrParseNode) {
-            StrParseNode front = (StrParseNode) head;
+
             //Do not add an empty string node
-            if (front.getValue().byteLength() == 0) {
-                head = createDStrNode(
-                        head.getPosition(),
-                        front.isFrozen());
+            if (((StrParseNode) head).getValue().byteLength() == 0) {
+                head = createDStrNode(head.getPosition());
             } else {
-                head = createDStrNode(head.getPosition(), front.isFrozen())
-                        .add(head);
+                head = createDStrNode(head.getPosition()).add(head);
             }
         }
         return ((DStrParseNode) head).add(tail);