+++ src/main/java/org/truffleruby/parser/parser/RubyParser.java
@@ -3133,7 +3133,7 @@ states[478] = (support, lexer, yyVal, yyVals, yyTop) -> {
     return yyVal;
 };
 states[480] = (support, lexer, yyVal, yyVals, yyTop) -> {
-    yyVal = ((ParseNode)yyVals[0+yyTop]) instanceof EvStrParseNode ? new DStrParseNode(((ParseNode)yyVals[0+yyTop]).getPosition(), lexer.getEncoding(), support.getConfiguration().isFrozenStringLiteral()).add(((ParseNode)yyVals[0+yyTop])) : ((ParseNode)yyVals[0+yyTop]);
+    yyVal = ((ParseNode)yyVals[0+yyTop]) instanceof EvStrParseNode ? new DStrParseNode(((ParseNode)yyVals[0+yyTop]).getPosition(), lexer.getEncoding()).add(((ParseNode)yyVals[0+yyTop])) : ((ParseNode)yyVals[0+yyTop]);
     /*
     NODE *node = $1;
     if (!node) {
@@ -3195,7 +3195,7 @@ states[488] = (support, lexer, yyVal, yyVals, yyTop) -> {
     return yyVal;
 };
 states[489] = (support, lexer, yyVal, yyVals, yyTop) -> {
-    yyVal = ((ListParseNode)yyVals[-2+yyTop]).add(((ParseNode)yyVals[-1+yyTop]) instanceof EvStrParseNode ? new DStrParseNode(((ListParseNode)yyVals[-2+yyTop]).getPosition(), lexer.getEncoding(), false).add(((ParseNode)yyVals[-1+yyTop])) : ((ParseNode)yyVals[-1+yyTop]));
+    yyVal = ((ListParseNode)yyVals[-2+yyTop]).add(((ParseNode)yyVals[-1+yyTop]) instanceof EvStrParseNode ? new DStrParseNode(((ListParseNode)yyVals[-2+yyTop]).getPosition(), lexer.getEncoding()).add(((ParseNode)yyVals[-1+yyTop])) : ((ParseNode)yyVals[-1+yyTop]));
     return yyVal;
 };
 states[490] = (support, lexer, yyVal, yyVals, yyTop) -> {
@@ -3878,7 +3878,7 @@ states[638] = (support, lexer, yyVal, yyVals, yyTop) -> {
 };
 states[639] = (support, lexer, yyVal, yyVals, yyTop) -> {
     if (((ParseNode)yyVals[-2+yyTop]) instanceof StrParseNode) {
-        DStrParseNode dnode = new DStrParseNode(support.getPosition(((ParseNode)yyVals[-2+yyTop])), lexer.getEncoding(), false);
+        DStrParseNode dnode = new DStrParseNode(support.getPosition(((ParseNode)yyVals[-2+yyTop])), lexer.getEncoding());
         dnode.add(((ParseNode)yyVals[-2+yyTop]));
         yyVal = support.createKeyValue(new DSymbolParseNode(support.getPosition(((ParseNode)yyVals[-2+yyTop])), dnode), ((ParseNode)yyVals[0+yyTop]));
     } else if (((ParseNode)yyVals[-2+yyTop]) instanceof DStrParseNode) {